refactor(config_compiler): seperate out part of the implementation to be used by plugins
